% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanfolder.R
\name{cleanfolder}
\alias{cleanfolder}
\title{Nettoyage}
\usage{
cleanfolder()
}
\arguments{
\item{rep}{= dossier à nettoyer

#' @author Nolwenn}
}
\description{
Suppression des fichiers inutiles
}
\examples{
cleanfolder()

}
