% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoKrigeageFonc.R
\name{CoKrigeageFonc}
\alias{CoKrigeageFonc}
\title{Fonction Krigeage avec peuplement}
\usage{
CoKrigeageFonc(grd, shpPlac, shp, idvar = "VcHa")
}
\arguments{
\item{grd}{= grid}

\item{shpPlac}{= placettes au format sf POINT}

\item{shp}{= fichier géoréférencé au format sf POLYGON correspondant aux peuplements ou aux
unités de gestion}

\item{idvar}{= variable Y retenue}
}
\description{
Fonction krigeage pour conversion des données ponctuelles. Prise en compte des peuplements
}
\examples{
k <- CoKrigeageFonc(grd, Plac, Peuplt)

}
\author{
Bruciamacchie Max
}
