% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IlotDataImport.R
\name{IlotDataImport}
\alias{IlotDataImport}
\title{Liste et import les couches vecteurs necessaires a l'iloscope}
\usage{
IlotDataImport(projet = NULL)
}
\arguments{
\item{projet}{= nom de la table en sortie}
}
\description{
Imports des différents fichiers vecteurs constitutifs d'un projet
}
\examples{
IlotDataImport("MotteServolex")

}
\author{
Bruciamacchie Max
}
