% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IlotKrigeage.R
\name{IlotKrigeage}
\alias{IlotKrigeage}
\title{Krigeage des valeurs placettes}
\usage{
IlotKrigeage(rep, shp, pas = 25)
}
\arguments{
\item{rep}{= dossier en cours}

\item{shp}{= perimètre}

\item{pas}{= résolution}
}
\description{
Conversion données ponctuelles en données surfaciques
}
\examples{
IlotKrigeage(perim)

}
\author{
Bruciamacchie Max
}
