% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IlotPrintSession.R
\name{IlotPrintSession}
\alias{IlotPrintSession}
\title{Edition par équipe}
\usage{
IlotPrintSession(rep)
}
\arguments{
\item{rep}{= dossier en cours}
}
\description{
La fonction permet d'analyser le choix de chaque équipe (format .pdf)
}
\examples{
IlotPrintEquipe(rep)

}
\author{
Bruciamacchie Max
}
