% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrigeageFonc.R
\name{KrigeageFonc}
\alias{KrigeageFonc}
\title{Fonction Krigeage}
\usage{
KrigeageFonc(grd, shpPlac, idvar = 2, pas = 50)
}
\description{
Fonction krigeage pour conversion données ponctuelles
}
\examples{
k <- IlotKrigeage()

}
\author{
Bruciamacchie Max
}
