% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjetChoisir.R
\name{ProjetChoisir}
\alias{ProjetChoisir}
\title{Iloscope : Choix du projet (c'est pas terrible)}
\usage{
ProjetChoisir()
}
\description{
Choix du projet en tour
}
\examples{
ProjetChoisir()

}
\author{
Bruciamacchie Max
}
