% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjetCreer.R
\name{ProjetCreer}
\alias{ProjetCreer}
\title{Iloscope : creer un nouveau projet projet}
\usage{
ProjetCreer(Nom)
}
\arguments{
\item{projet}{= nom de la table en sortie}
}
\description{
Imports des différents fichiers vecteurs constitutifs d'un projet
}
\examples{
ProjetCreer("Test")

}
\author{
Bruciamacchie Max
}
