% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifData.R
\name{VerifData}
\alias{VerifData}
\title{Edition par équipe}
\usage{
VerifData(rep)
}
\arguments{
\item{rep}{= dossier en cours}
}
\description{
La fonction permet d'analyser le choix de chaque équipe (format .pdf)
}
\examples{
IlotPrintEquipe(rep)

}
\author{
Bruciamacchie Max
}
