% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IlotKrigeage.R
\name{IlotKrigeage}
\alias{IlotKrigeage}
\title{Krigeage des valeurs placettes}
\usage{
IlotKrigeage(pas = 50)
}
\description{
Conversion données ponctuelles en données surfaciques
}
\examples{
k <- IlotKrigeage()
plot(k["var1.pred"])

}
\author{
Bruciamacchie Max
}
